Allow excluding based on a provided collection
