+++ platforms/core-configuration/model-core/src/main/java/org/gradle/api/internal/provider/CollectionSupplier.java
@@ -29,4 +29,6 @@ interface CollectionSupplier<T, C extends Collection<? extends T>> extends Value
     CollectionSupplier<T, C> keep(Predicate<T> filter);
 
     CollectionSupplier<T, C> pruned();
+
+    CollectionSupplier<T,C> minus(Collector<T> collector);
 }
